refactor: reduce helpers
