+++ src/helpers/parseBlinksFromContent.ts
@@ -1,8 +1,45 @@
 import { uniqBy } from 'lodash-es';
 
 import { CONTENT_SOLANA_BLINK_REGEX } from '@/constants/regexp.js';
-import { decodeActionUrl } from '@/helpers/decodeActionUrl.js';
 import { matchUrls } from '@/helpers/matchUrls.js';
+import { parseURL } from '@/helpers/parseURL.js';
+
+type ActionUrl =
+    | {
+          isBlink: true;
+          decodedActionUrl: string;
+          url: string;
+      }
+    | {
+          isBlink: false;
+          url: string;
+      };
+
+function decodeActionUrl(url: string): ActionUrl {
+    const parsedUrl = parseURL(url);
+    if (!parsedUrl) return { isBlink: false, url };
+
+    const actionUrl = parsedUrl.searchParams.get('action');
+    if (!actionUrl) {
+        return { isBlink: false, url };
+    }
+    const urlDecodedActionUrl = decodeURIComponent(actionUrl);
+
+    const solanaActionPrefix = /^(solana-action:|solana:)/;
+    if (!solanaActionPrefix.test(urlDecodedActionUrl)) {
+        return { isBlink: false, url };
+    }
+    const decodedActionUrl = urlDecodedActionUrl.replace(solanaActionPrefix, '');
+
+    const parsedActionUrl = parseURL(decodedActionUrl);
+    if (!parsedActionUrl) return { isBlink: false, url };
+
+    return {
+        url,
+        isBlink: true,
+        decodedActionUrl: parsedActionUrl.toString(),
+    };
+}
 
 export function parseBlinksFromContent(
     content: string,
