+++ src/providers/blink/Loader.ts
@@ -5,7 +5,6 @@ import { anySignal } from '@/helpers/anySignal.js';
 import { fetchCachedJSON, fetchJSON } from '@/helpers/fetchJSON.js';
 import { parseURL } from '@/helpers/parseURL.js';
 import { requestIdleCallbackAsync } from '@/helpers/requestIdleCallbackAsync.js';
-import { resolveActionJson } from '@/helpers/resolveActionJson.js';
 import { BaseLoader } from '@/providers/base/Loader.js';
 import type {
     Action,
@@ -14,6 +13,30 @@ import type {
     ActionsSpecGetResponse,
     Parameter,
 } from '@/providers/types/Blink.js';
+
+function resolveActionJson(url: string, actions: ActionRuleResponse) {
+    const urlObj = parseURL(url);
+    if (!urlObj) throw new Error('Invalid blink');
+    const paths = urlObj.pathname.split('/');
+    for (const rule of actions.rules) {
+        const pathPatterns = rule.pathPattern.split('/');
+        for (let i = 0; i < pathPatterns.length; i += 1) {
+            const pathPattern = pathPatterns[i];
+            if (pathPattern === '**') {
+                pathPatterns[i] = paths.slice(i).join('/');
+                continue;
+            }
+            if (pathPattern === '*') {
+                pathPatterns[i] = paths[i];
+                continue;
+            }
+            if (pathPattern !== paths[i]) break;
+        }
+        const newPath = pathPatterns.join('/');
+        if (newPath !== rule.pathPattern) return rule.apiPath.replace(rule.pathPattern, pathPatterns.join('/'));
+    }
+    return null;
+}
 
 function createActionComponent(label: string, href: string, parameters?: [Parameter]): ActionComponent {
     return {
